rename button folders xButton to x
